[INV-47] Merged Vue files into server/
**Changed:**
- server/build.sh changed Sass and Vue compilation to point to the new
location.
- /server/serverRun.py
    - added commented line that serves the Vue homepage
over the Flask one.
    - Flask now serves from /public/ not /static/
- Templated Flask HTML files now have img and style paths without
prepended "/static/" or "/public/"
