+++ server/public/index.html
@@ -5,7 +5,8 @@
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width,initial-scale=1.0">
     <link rel="icon" href="<%= BASE_URL %>favicon.ico">
-    <title>vue</title>
+    <link rel="stylesheet" type="text/css" href="<%= BASE_URL %>css/styles.css">
+    <title>Vue</title>
   </head>
   <body>
     <noscript>
