+++ server/src/assets/logo.png
