+++ server/src/views/About.vue
