+++ server/templates/pages/login.html
@@ -3,7 +3,7 @@
 {% block content %}
 <div class="flex-vert modal-content">
     <div class="user-img">
-        <img src="static/img/face.png">
+        <img src="/img/face.png">
     </div>
     <div class="form-input flex-vert">
         <form id="login form" action="/login" method="post">
@@ -29,6 +29,6 @@
     </div>
 </div>
 
-<script src="/static/js/field_validators.js"></script>
-<script src="/static/js/login.js"></script>
+<script src="/js/field_validators.js"></script>
+<script src="/js/login.js"></script>
 {% endblock %}
