+++ server/templates/pages/signup.html
@@ -2,7 +2,7 @@
 {% block content %}
 <div class="modal-content flex-vert">
     <div class="user-img">
-        <img src="static/img/add_user_round.png">
+        <img src="/img/add_user_round.png">
     </div>
 
     <div class="form-input flex-vert">
@@ -27,7 +27,7 @@
         <a href="/login.html">Log In</a>
     </div>
 </div>
-<script src="/static/js/field_validators.js"></script>
-<script src="/static/js/signup.js"></script>
+<script src="/js/field_validators.js"></script>
+<script src="/js/signup.js"></script>
 {% endblock %}
